scripts/bodacc/requirements.txt
requests==2.32.3
feedparser==6.0.10
python-dateutil==2.9.0.post0